+++ app/src/main/java/com/toyou/toyouandroid/presentation/fragment/social/SendFinalFragment.kt
@@ -1,6 +1,8 @@
-package com.toyou.toyouandroid.ui.social
+package com.toyou.toyouandroid.presentation.fragment.social
 
 import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -8,13 +10,13 @@ import androidx.fragment.app.Fragment
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
 import com.toyou.toyouandroid.R
-import com.toyou.toyouandroid.databinding.FragmentQuestionContentBinding
 import com.toyou.toyouandroid.databinding.FragmentSendBinding
+import com.toyou.toyouandroid.databinding.FragmentSendFinalBinding
 import com.toyou.toyouandroid.presentation.base.MainActivity
 
-class SendFragment: Fragment() {
-    private var _binding : FragmentSendBinding? = null
-    private val binding : FragmentSendBinding get() = requireNotNull(_binding){"널"}
+class SendFinalFragment: Fragment() {
+    private var _binding : FragmentSendFinalBinding? = null
+    private val binding : FragmentSendFinalBinding get() = requireNotNull(_binding){"널"}
 
     private lateinit var navController: NavController
 
@@ -27,7 +29,7 @@ class SendFragment: Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentSendBinding.inflate(inflater, container, false)
+        _binding = FragmentSendFinalBinding.inflate(inflater, container, false)
 
 
         return binding.root
@@ -38,12 +40,11 @@ class SendFragment: Fragment() {
         super.onViewCreated(view, savedInstanceState)
         navController = Navigation.findNavController(view)
 
-        binding.nextBtn.setOnClickListener {
-            navController.navigate(R.id.action_sendFragment_to_sendFinalFragment)
-        }
-        binding.backBtn.setOnClickListener {
-            navController.popBackStack()
+        Handler(Looper.getMainLooper()).postDelayed({
+            val mainActivity = activity as MainActivity
+            mainActivity.hideBottomNavigation(false)
+            navController.navigate(R.id.action_sendFinalFragment_to_navigation_home)
+        }, 2000)
 
-        }
     }
 }
