+++ app/src/main/java/com/toyou/toyouandroid/presentation/fragment/social/adapter/SocialRVAdapter.kt
@@ -1,11 +1,10 @@
-package com.toyou.toyouandroid.ui.social.adapter
+package com.toyou.toyouandroid.presentation.fragment.social.adapter
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageButton
 import android.widget.TextView
-import androidx.appcompat.view.menu.MenuView.ItemView
 import androidx.recyclerview.widget.RecyclerView
 import com.toyou.toyouandroid.R
 import com.toyou.toyouandroid.model.FriendListModel
@@ -21,12 +20,12 @@ class SocialRVAdapter(private val onItemClick: (Int) -> Unit) : RecyclerView.Ada
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int
-    ): SocialRVAdapter.SocialViewHolder {
+    ): SocialViewHolder {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.item_friend_rv, parent, false)
         return SocialViewHolder(view, onItemClick)
     }
 
-    override fun onBindViewHolder(holder: SocialRVAdapter.SocialViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: SocialViewHolder, position: Int) {
         holder.bind(friendList[position])
     }
